Restructure analyze module
