+++ crates/engine/analysis/src/api/query/symbols/document.rs
@@ -1,31 +1,28 @@
-//! Symbol queries for editor outlines and workspace-wide search.
-//!
-//! These APIs deliberately stay transport-neutral: they expose LSP-shaped facts, but do not depend
-//! on LSP enums or protocol structs. The future server can map this data into document/workspace
-//! symbol responses without teaching the analysis layer about client capabilities.
+//! Document symbol query for editor outlines.
 
 use anyhow::Result;
 use rg_body_ir::{BodyData, BodyImplData, BodyItemData};
-use rg_def_map::{LocalDefRef, ModuleData, ModuleId, ModuleOrigin, ModuleRef, TargetRef};
+use rg_def_map::{LocalDefRef, TargetRef};
 use rg_parse::{FileId, Span};
 use rg_semantic_ir::{
-    AssocItemId, ConstData, ConstRef, EnumData, FunctionData, FunctionRef, ImplData, ItemOwner,
-    StaticData, StructData, TraitData, TypeAliasData, TypeAliasRef, TypeDefRef, UnionData,
+    AssocItemId, ConstData, ConstRef, EnumData, FunctionData, FunctionRef, ItemOwner, StaticData,
+    StructData, TypeAliasData, TypeAliasRef, UnionData,
 };
 
-use super::{
-    Analysis,
-    data::{DocumentSymbol, SymbolKind, WorkspaceSymbol},
+use super::shared;
+use crate::{
+    api::Analysis,
+    model::{DocumentSymbol, SymbolKind},
 };
 
-pub(super) struct SymbolCollector<'a, 'db>(&'a Analysis<'db>);
+pub(crate) struct DocumentSymbolCollector<'a, 'db>(&'a Analysis<'db>);
 
-impl<'a, 'db> SymbolCollector<'a, 'db> {
-    pub(super) fn new(analysis: &'a Analysis<'db>) -> Self {
+impl<'a, 'db> DocumentSymbolCollector<'a, 'db> {
+    pub(crate) fn new(analysis: &'a Analysis<'db>) -> Self {
         Self(analysis)
     }
 
-    pub(super) fn document_symbols(
+    pub(crate) fn document_symbols(
         &self,
         target: TargetRef,
         file_id: FileId,
@@ -51,35 +48,6 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
         Ok(symbols)
     }
 
-    pub(super) fn workspace_symbols(&self, query: &str) -> Result<Vec<WorkspaceSymbol>> {
-        let query = WorkspaceSymbolQuery::new(query);
-        let mut symbols = Vec::new();
-
-        for (target, _) in self.0.semantic_ir.materialize_included_target_irs()? {
-            self.push_module_workspace_symbols(target, &query, &mut symbols)?;
-            self.push_nominal_workspace_symbols(target, &query, &mut symbols)?;
-            self.push_trait_workspace_symbols(target, &query, &mut symbols)?;
-            self.push_impl_workspace_symbols(target, &query, &mut symbols)?;
-            self.push_function_workspace_symbols(target, &query, &mut symbols)?;
-            self.push_type_alias_workspace_symbols(target, &query, &mut symbols)?;
-            self.push_const_workspace_symbols(target, &query, &mut symbols)?;
-            self.push_static_workspace_symbols(target, &query, &mut symbols)?;
-        }
-
-        symbols.sort_by_key(|symbol| {
-            (
-                symbol.name.to_lowercase(),
-                symbol.kind,
-                symbol.container_name.clone(),
-                symbol.target.package.0,
-                symbol.target.target.0,
-                symbol.file_id.0,
-                symbol.span.map(|span| span.text.start),
-            )
-        });
-        Ok(symbols)
-    }
-
     fn push_module_document_symbols(
         &self,
         target: TargetRef,
@@ -90,7 +58,7 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
             let Some(name) = data.name.as_ref().map(ToString::to_string) else {
                 continue;
             };
-            let Some(source) = Self::module_declaration_source(data) else {
+            let Some(source) = shared::module_declaration_source(data) else {
                 continue;
             };
             if source.file_id != file_id {
@@ -156,7 +124,7 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
                 .map(|field| {
                     Self::field_document_symbol(
                         file_id,
-                        Self::field_label(field.key.as_ref().map(|key| key.declaration_label())),
+                        shared::field_label(field.key.as_ref().map(|key| key.declaration_label())),
                         field.span,
                     )
                 })
@@ -184,7 +152,7 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
                 .map(|field| {
                     Self::field_document_symbol(
                         file_id,
-                        Self::field_label(field.key.as_ref().map(|key| key.declaration_label())),
+                        shared::field_label(field.key.as_ref().map(|key| key.declaration_label())),
                         field.span,
                     )
                 })
@@ -222,7 +190,7 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
                         .map(|field| {
                             Self::field_document_symbol(
                                 file_id,
-                                Self::field_label(
+                                shared::field_label(
                                     field.key.as_ref().map(|key| key.declaration_label()),
                                 ),
                                 field.span,
@@ -273,7 +241,7 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
             }
 
             symbols.push(DocumentSymbol {
-                name: Self::impl_label(data),
+                name: shared::impl_label(data),
                 kind: SymbolKind::Impl,
                 file_id,
                 span: local_impl.span,
@@ -505,7 +473,7 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
                 .map(|field| {
                     Self::field_document_symbol(
                         file_id,
-                        Self::field_label(field.key.as_ref().map(|key| key.declaration_label())),
+                        shared::field_label(field.key.as_ref().map(|key| key.declaration_label())),
                         field.span,
                     )
                 })
@@ -519,7 +487,7 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
         impl_data: &BodyImplData,
     ) -> DocumentSymbol {
         DocumentSymbol {
-            name: Self::body_impl_label(impl_data),
+            name: shared::body_impl_label(impl_data),
             kind: SymbolKind::Impl,
             file_id: impl_data.source.file_id,
             span: impl_data.source.span,
@@ -563,415 +531,6 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
         None
     }
 
-    fn push_nominal_workspace_symbols(
-        &self,
-        target: TargetRef,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for (ty, data) in self.0.semantic_ir.structs(target)? {
-            self.push_local_def_workspace_symbol(
-                target,
-                data.local_def,
-                SymbolKind::Struct,
-                &data.name,
-                None,
-                query,
-                symbols,
-            )?;
-            self.push_field_workspace_symbols(ty, &data.name, query, symbols)?;
-        }
-
-        for (ty, data) in self.0.semantic_ir.unions(target)? {
-            self.push_local_def_workspace_symbol(
-                target,
-                data.local_def,
-                SymbolKind::Union,
-                &data.name,
-                None,
-                query,
-                symbols,
-            )?;
-            self.push_field_workspace_symbols(ty, &data.name, query, symbols)?;
-        }
-
-        for (_, data) in self.0.semantic_ir.enums(target)? {
-            self.push_local_def_workspace_symbol(
-                target,
-                data.local_def,
-                SymbolKind::Enum,
-                &data.name,
-                None,
-                query,
-                symbols,
-            )?;
-            for variant in &data.variants {
-                self.push_workspace_symbol(
-                    WorkspaceSymbolInput {
-                        target,
-                        name: variant.name.to_string(),
-                        kind: SymbolKind::EnumVariant,
-                        file_id: data.source.file_id,
-                        span: Some(variant.name_span),
-                        container_name: Some(data.name.to_string()),
-                    },
-                    query,
-                    symbols,
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    fn push_module_workspace_symbols(
-        &self,
-        target: TargetRef,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for (module_ref, data) in self.0.def_map.modules(target)? {
-            let Some(name) = data.name.as_ref().map(ToString::to_string) else {
-                continue;
-            };
-            let Some(source) = Self::module_declaration_source(data) else {
-                continue;
-            };
-
-            self.push_workspace_symbol(
-                WorkspaceSymbolInput {
-                    target,
-                    name,
-                    kind: SymbolKind::Module,
-                    file_id: source.file_id,
-                    span: Some(source.selection_span),
-                    container_name: self.module_container_name(module_ref)?,
-                },
-                query,
-                symbols,
-            );
-        }
-
-        Ok(())
-    }
-
-    fn push_trait_workspace_symbols(
-        &self,
-        target: TargetRef,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for (_, data) in self.0.semantic_ir.traits(target)? {
-            self.push_local_def_workspace_symbol(
-                target,
-                data.local_def,
-                SymbolKind::Trait,
-                &data.name,
-                None,
-                query,
-                symbols,
-            )?;
-            self.push_assoc_item_workspace_symbols(
-                target,
-                &data.items,
-                &Self::trait_label(data),
-                query,
-                symbols,
-            )?;
-        }
-
-        Ok(())
-    }
-
-    fn push_impl_workspace_symbols(
-        &self,
-        target: TargetRef,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for (impl_ref, data) in self.0.semantic_ir.impls(target)? {
-            self.push_assoc_item_workspace_symbols(
-                impl_ref.target,
-                &data.items,
-                &Self::impl_label(data),
-                query,
-                symbols,
-            )?;
-        }
-
-        Ok(())
-    }
-
-    fn push_function_workspace_symbols(
-        &self,
-        target: TargetRef,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for (_, data) in self.0.semantic_ir.functions(target)? {
-            if !matches!(data.owner, ItemOwner::Module(_)) {
-                continue;
-            }
-            self.push_workspace_symbol(
-                WorkspaceSymbolInput {
-                    target,
-                    name: data.name.to_string(),
-                    kind: SymbolKind::Function,
-                    file_id: data.source.file_id,
-                    span: Some(data.name_span.unwrap_or(data.span)),
-                    container_name: None,
-                },
-                query,
-                symbols,
-            );
-        }
-
-        Ok(())
-    }
-
-    fn push_type_alias_workspace_symbols(
-        &self,
-        target: TargetRef,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for (_, data) in self.0.semantic_ir.type_aliases(target)? {
-            if !matches!(data.owner, ItemOwner::Module(_)) {
-                continue;
-            }
-            self.push_type_alias_workspace_symbol(target, data, None, query, symbols);
-        }
-
-        Ok(())
-    }
-
-    fn push_const_workspace_symbols(
-        &self,
-        target: TargetRef,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for (_, data) in self.0.semantic_ir.consts(target)? {
-            if !matches!(data.owner, ItemOwner::Module(_)) {
-                continue;
-            }
-            self.push_const_workspace_symbol(target, data, None, query, symbols);
-        }
-
-        Ok(())
-    }
-
-    fn push_static_workspace_symbols(
-        &self,
-        target: TargetRef,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for (_, data) in self.0.semantic_ir.statics(target)? {
-            self.push_local_def_workspace_symbol(
-                target,
-                data.local_def,
-                SymbolKind::Static,
-                &data.name,
-                None,
-                query,
-                symbols,
-            )?;
-        }
-
-        Ok(())
-    }
-
-    fn push_assoc_item_workspace_symbols(
-        &self,
-        target: TargetRef,
-        items: &[AssocItemId],
-        container_name: &str,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for item in items {
-            match item {
-                AssocItemId::Function(id) => {
-                    let function_ref = FunctionRef { target, id: *id };
-                    let Some(data) = self.0.semantic_ir.function_data(function_ref)? else {
-                        continue;
-                    };
-                    self.push_workspace_symbol(
-                        WorkspaceSymbolInput {
-                            target,
-                            name: data.name.to_string(),
-                            kind: SymbolKind::Method,
-                            file_id: data.source.file_id,
-                            span: Some(data.name_span.unwrap_or(data.span)),
-                            container_name: Some(container_name.to_string()),
-                        },
-                        query,
-                        symbols,
-                    );
-                }
-                AssocItemId::TypeAlias(id) => {
-                    let type_alias_ref = TypeAliasRef { target, id: *id };
-                    let Some(data) = self.0.semantic_ir.type_alias_data(type_alias_ref)? else {
-                        continue;
-                    };
-                    self.push_type_alias_workspace_symbol(
-                        target,
-                        data,
-                        Some(container_name.to_string()),
-                        query,
-                        symbols,
-                    );
-                }
-                AssocItemId::Const(id) => {
-                    let const_ref = ConstRef { target, id: *id };
-                    let Some(data) = self.0.semantic_ir.const_data(const_ref)? else {
-                        continue;
-                    };
-                    self.push_const_workspace_symbol(
-                        target,
-                        data,
-                        Some(container_name.to_string()),
-                        query,
-                        symbols,
-                    );
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn push_field_workspace_symbols(
-        &self,
-        ty: TypeDefRef,
-        container_name: &str,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        for field_ref in self.0.semantic_ir.fields_for_type(ty)? {
-            let Some(field_data) = self.0.semantic_ir.field_data(field_ref)? else {
-                continue;
-            };
-            let name = Self::field_label(
-                field_data
-                    .field
-                    .key
-                    .as_ref()
-                    .map(|key| key.declaration_label()),
-            );
-            self.push_workspace_symbol(
-                WorkspaceSymbolInput {
-                    target: ty.target,
-                    name,
-                    kind: SymbolKind::Field,
-                    file_id: field_data.file_id,
-                    span: Some(field_data.field.span),
-                    container_name: Some(container_name.to_string()),
-                },
-                query,
-                symbols,
-            );
-        }
-
-        Ok(())
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn push_local_def_workspace_symbol(
-        &self,
-        target: TargetRef,
-        local_def: LocalDefRef,
-        kind: SymbolKind,
-        name: &str,
-        container_name: Option<String>,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) -> Result<()> {
-        let Some(data) = self.0.def_map.local_def(local_def)? else {
-            return Ok(());
-        };
-
-        self.push_workspace_symbol(
-            WorkspaceSymbolInput {
-                target,
-                name: name.to_string(),
-                kind,
-                file_id: data.file_id,
-                span: Some(data.name_span.unwrap_or(data.span)),
-                container_name,
-            },
-            query,
-            symbols,
-        );
-
-        Ok(())
-    }
-
-    fn push_type_alias_workspace_symbol(
-        &self,
-        target: TargetRef,
-        data: &TypeAliasData,
-        container_name: Option<String>,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) {
-        self.push_workspace_symbol(
-            WorkspaceSymbolInput {
-                target,
-                name: data.name.to_string(),
-                kind: SymbolKind::TypeAlias,
-                file_id: data.source.file_id,
-                span: Some(data.name_span.unwrap_or(data.span)),
-                container_name,
-            },
-            query,
-            symbols,
-        );
-    }
-
-    fn push_const_workspace_symbol(
-        &self,
-        target: TargetRef,
-        data: &ConstData,
-        container_name: Option<String>,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) {
-        self.push_workspace_symbol(
-            WorkspaceSymbolInput {
-                target,
-                name: data.name.to_string(),
-                kind: SymbolKind::Const,
-                file_id: data.source.file_id,
-                span: Some(data.name_span.unwrap_or(data.span)),
-                container_name,
-            },
-            query,
-            symbols,
-        );
-    }
-
-    fn push_workspace_symbol(
-        &self,
-        input: WorkspaceSymbolInput,
-        query: &WorkspaceSymbolQuery,
-        symbols: &mut Vec<WorkspaceSymbol>,
-    ) {
-        if !query.matches(&input.name) {
-            return;
-        }
-
-        symbols.push(WorkspaceSymbol {
-            target: input.target,
-            name: input.name,
-            kind: input.kind,
-            file_id: input.file_id,
-            span: input.span,
-            container_name: input.container_name,
-        });
-    }
-
     fn field_document_symbol(file_id: FileId, name: String, span: Span) -> DocumentSymbol {
         DocumentSymbol {
             name,
@@ -987,7 +546,7 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
         &self,
         local_def: LocalDefRef,
         file_id: FileId,
-    ) -> Result<Option<SymbolSource>> {
+    ) -> Result<Option<shared::SymbolSource>> {
         let Some(data) = self.0.def_map.local_def(local_def)? else {
             return Ok(None);
         };
@@ -995,62 +554,10 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
             return Ok(None);
         }
 
-        Ok(Some(SymbolSource {
+        Ok(Some(shared::SymbolSource {
             span: data.span,
             selection_span: data.name_span.unwrap_or(data.span),
         }))
-    }
-
-    fn module_declaration_source(module: &ModuleData) -> Option<SymbolSourceWithFile> {
-        let (file_id, span) = match module.origin {
-            ModuleOrigin::Root { .. } => return None,
-            ModuleOrigin::Inline {
-                declaration_file,
-                declaration_span,
-            }
-            | ModuleOrigin::OutOfLine {
-                declaration_file,
-                declaration_span,
-                ..
-            } => (declaration_file, declaration_span),
-        };
-
-        Some(SymbolSourceWithFile {
-            file_id,
-            span,
-            selection_span: module.name_span.unwrap_or(span),
-        })
-    }
-
-    fn module_container_name(&self, module_ref: ModuleRef) -> Result<Option<String>> {
-        let Some(module) = self.0.def_map.module(module_ref)? else {
-            return Ok(None);
-        };
-        let Some(parent) = module.parent else {
-            return Ok(None);
-        };
-        let path = self.module_path(module_ref.target, parent)?;
-
-        Ok((!path.is_empty()).then_some(path))
-    }
-
-    fn module_path(&self, target: TargetRef, module: ModuleId) -> Result<String> {
-        let Some(data) = self.0.def_map.module(ModuleRef { target, module })? else {
-            return Ok(String::new());
-        };
-        let Some(name) = &data.name else {
-            return Ok(String::new());
-        };
-        let Some(parent) = data.parent else {
-            return Ok(name.to_string());
-        };
-
-        let parent_path = self.module_path(target, parent)?;
-        if parent_path.is_empty() {
-            Ok(name.to_string())
-        } else {
-            Ok(format!("{parent_path}::{name}"))
-        }
     }
 
     fn nest_module_document_symbols(symbols: Vec<DocumentSymbol>) -> Vec<DocumentSymbol> {
@@ -1129,63 +636,5 @@ impl<'a, 'db> SymbolCollector<'a, 'db> {
                 symbol.name.clone(),
             )
         });
-    }
-
-    fn field_label(name: Option<String>) -> String {
-        name.unwrap_or_else(|| "<unsupported>".to_string())
-    }
-
-    fn trait_label(data: &TraitData) -> String {
-        format!("trait {}", data.name)
-    }
-
-    fn impl_label(data: &ImplData) -> String {
-        match &data.trait_ref {
-            Some(trait_ref) => format!("impl {trait_ref} for {}", data.self_ty),
-            None => format!("impl {}", data.self_ty),
-        }
-    }
-
-    fn body_impl_label(data: &BodyImplData) -> String {
-        match &data.trait_ref {
-            Some(trait_ref) => format!("impl {trait_ref} for {}", data.self_ty),
-            None => format!("impl {}", data.self_ty),
-        }
-    }
-}
-
-struct SymbolSource {
-    span: Span,
-    selection_span: Span,
-}
-
-struct SymbolSourceWithFile {
-    file_id: FileId,
-    span: Span,
-    selection_span: Span,
-}
-
-struct WorkspaceSymbolInput {
-    target: TargetRef,
-    name: String,
-    kind: SymbolKind,
-    file_id: FileId,
-    span: Option<Span>,
-    container_name: Option<String>,
-}
-
-struct WorkspaceSymbolQuery {
-    needle: String,
-}
-
-impl WorkspaceSymbolQuery {
-    fn new(query: &str) -> Self {
-        Self {
-            needle: query.to_lowercase(),
-        }
-    }
-
-    fn matches(&self, name: &str) -> bool {
-        self.needle.is_empty() || name.to_lowercase().contains(&self.needle)
     }
 }
